+++ src/client/java/tempeststudios/quickstacknearby/QuickStackRulesButtonScreen.java
@@ -1,0 +1,5 @@
+package tempeststudios.quickstacknearby;
+
+public interface QuickStackRulesButtonScreen {
+    boolean quickstacknearby$openRulesFromButton(double mouseX, double mouseY, int button);
+}